w1retap-0.0.11.tar.gz
sha512sum c656192a5a3e832ab8b7e6a87bf8163f0e81b57aadeefc9ebf1344e38714e3c1a6ef41cfcebef1452535327bd1a19b1a220228d90dd3076d195558c277294d11

+++ src/libusblinux300/atod20.c
@@ -253,7 +253,7 @@ int ReadAtoDResults(int portnum, int try_overdrive, uchar *SerialNum,
                          float *prslt, uchar *ctrl)
 {
    uchar i, send_cnt=0;
-   ulong templong;
+   unsigned int templong;
    uchar rt=FALSE;
    uchar send_block[30];
    ushort lastcrc16;
